Add files via upload
In this project, a generic music player system using python program with certain features is implemented. The goal is to use data structure and perform certain operations on the music player where songs are added. Here the implementation consist basic use of OOP and fundamental concepts like sorting, insertion, deletion of the songs from the created playlist.